Advanced search implemented
